+++ nav2_behaviors/CMakeLists.txt
@@ -55,14 +55,6 @@ ament_target_dependencies(nav2_spin_behavior
   ${dependencies}
 )
 
-add_library(nav2_backup_behavior SHARED
-  plugins/back_up.cpp
-)
-
-ament_target_dependencies(nav2_backup_behavior
-  ${dependencies}
-)
-
 add_library(nav2_wait_behavior SHARED
 plugins/wait.cpp
 )
@@ -104,7 +96,6 @@ ament_target_dependencies(${executable_name}
 rclcpp_components_register_nodes(${library_name} "behavior_server::BehaviorServer")
 
 install(TARGETS ${library_name}
-                nav2_backup_behavior
                 nav2_spin_behavior
                 nav2_wait_behavior
                 nav2_drive_on_heading_behavior
@@ -134,7 +125,6 @@ endif()
 
 ament_export_include_directories(include)
 ament_export_libraries(${library_name}
-  nav2_backup_behavior
   nav2_spin_behavior
   nav2_wait_behavior
   nav2_drive_on_heading_behavior
